\name{index_raster_to_mask}
\alias{index_raster_to_mask}
\title{Internal function}
\usage{
  index_raster_to_mask(index_raster, nlayers)
}
\arguments{
  \item{index_raster}{TODO}

  \item{nlayers}{TODO}
}
\description{
  Internal function
}
\author{
  Jonathan A. Greenberg
}
\seealso{
  \code{\link[base]{which.min}},
  \code{\link[base]{which.max}}
}
\keyword{brick/stack}
\keyword{format}

