\name{ncdf_gaussian_grid_to_raster}
\alias{ncdf_gaussian_grid_to_raster}
\title{Create Raster from NCDF}
\usage{
  ncdf_gaussian_grid_to_raster(ncdf_gaussian_grid_fname,
    reference)
}
\arguments{
  \item{ncdf_gaussian_grid_fname}{TODO}

  \item{reference}{TODO}
}
\description{
  Creates a raster from an NCDF file
}
\author{
  Jonathan A. Greenberg
}
\seealso{
  TODO
}
\keyword{brick/stack}
\keyword{format}

