\name{windvectors_to_wind}
\alias{windvectors_to_wind}
\title{Calculate Wind Speed}
\usage{
  windvectors_to_wind(uwnd, vwnd, wnd_height)
}
\arguments{
  \item{uwnd}{TODO}

  \item{vwnd}{TODO}

  \item{wnd_height}{TODO}
}
\description{
  Takes two wind vectors and returns wind speed, correcting
  to 2m above the terrain if requested.
}
\author{
  Jonathan A. Greenberg
}
\seealso{
  TODO
}
\keyword{climate}

