\name{climstats-package}
\alias{climstats-package}
\alias{climstats}
\docType{package}

\title{ climstats }

\description{
These functions are designed to prepare and summarize climate data time series (point based or raster) for use in, for instance, species distribution models.
}
\details{
\tabular{ll}{
Package: \tab climstats\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2010-10-13\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Jonathan Asher Greenberg

Maintainer: Jonathan Asher Greenberg <climstats@estarcion.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}


\keyword{ package }

