\name{prepare_climate_data}

\alias{prepare_climate_data}


\title{
This is the tile
}
\description{
TBD
}

\usage{
prepare_climate_data(raster_files, raster_source, zname, zvalue, proj)
}


\arguments{
  \item{raster_files}{vector of fileanames}
  \item{raster_source}{that}
  \item{zname}{this}
  \item{zvalue}{that}
  \item{proj}{CRS }
}

\details{
no details
}

\value{
}

\references{
none
}

\author{
Jonathan Greenberg
}

\note{
no notes
}

\seealso{
\code{\link[climstats]{apply_gains_offsets}} 
}

\examples{
x = 1:10
x * 2
}

\keyword{ spatial }
